backend/requirements.txt
uvicorn
fastapi
google-adk
google-generativeai
webrtcvad
setuptools
